Forms back-end
Split questions of submitted matrix field into individual radio fields.

Append 0's to form entry table names making it three digits.

Rename form_fields target_type_id column to reference_target_type_id and
target_id column to reference_target_id.

Rename form_fields table label column to title.

Remove form_fields timestamps.

Remove id column from form_fields and add column_name column. Primary
key becomes the column_name and form_id.

Change convention of field types saved in the database.

Create files table migration and set form entry columns for files as a
foreign key to the files table.

Group many-to-many relationships between forms and resources, specifically,
teams, programs, groups into a single polymorphic relationship.

Create resources for Forms.

Create form entry endpoint and return form data from there.

Create form entry route in the front-end.

+++ database/migrations/2018_09_14_000011_create_programs_table.php
@@ -17,7 +17,7 @@ class CreateProgramsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name', 32);
             $table->text('description')->nullable();
-            $table->integer('team_id')->unsigned();
+            $table->bigInteger('team_id')->unsigned();
             $table->integer('default_client_status_id')->unsigned();
             $table->integer('group_client_status_id')->unsigned()->nullable();
             $table->integer('case_client_status_id')->unsigned()->nullable();
